+++ chirho/explainable/handlers/__init__.py
@@ -1,4 +1,4 @@
-from .alternatives import random_intervention  # noqa: F401
-from .explanation import SearchForExplanation  # noqa: F401
+from .components import random_intervention  # noqa: F401
+from .components import undo_split  # noqa: F401
+from .explanation import SearchForExplanation, SplitSubsets  # noqa: F401
 from .preemptions import Preemptions  # noqa: F401
-from .split_subsets import SplitSubsets, undo_split  # noqa: F401
