+++ chirho/explainable/ops.py
@@ -1,11 +1,9 @@
 import functools
-from typing import Callable, Iterable, Optional, Tuple, TypeVar
+from typing import Optional, Tuple, TypeVar
 
 import pyro
-import torch
 
-from chirho.counterfactual.handlers.selection import get_factual_indices
-from chirho.indexed.ops import IndexSet, cond, cond_n, gather
+from chirho.indexed.ops import IndexSet, cond_n
 from chirho.interventional.ops import Intervention, intervene
 
 S = TypeVar("S")
@@ -42,37 +40,3 @@ def preempt(
         act_values[IndexSet(**{name: {i + 1}})] = intervene(obs, act, **kwargs)
 
     return cond_n(act_values, case, event_dim=kwargs.get("event_dim", 0))
-
-
-def consequent_differs(
-    antecedents: Iterable[str] = [], eps: float = -1e8, event_dim: int = 0
-) -> Callable[[T], torch.Tensor]:
-    """
-    A helper function for assessing whether values at a site differ from their observed values, assigning
-    `eps` if a value differs from its observed state and `0.0` otherwise.
-
-    :param antecedents: A list of names of upstream intervened sites to consider when assessing differences.
-    :param eps: A numerical value assigned if the values differ, defaults to -1e8.
-    :param event_dim: The event dimension of the value object.
-
-    :return: A callable which applied to a site value object (`consequent`), returns a tensor where each
-             element indicates whether the corresponding element of `consequent` differs from its factual value
-             (`eps` if there is a difference, `0.0` otherwise).
-    """
-
-    def _consequent_differs(consequent: T) -> torch.Tensor:
-        indices = IndexSet(
-            **{
-                name: ind
-                for name, ind in get_factual_indices().items()
-                if name in antecedents
-            }
-        )
-        not_eq: torch.Tensor = consequent != gather(
-            consequent, indices, event_dim=event_dim
-        )
-        for _ in range(event_dim):
-            not_eq = torch.all(not_eq, dim=-1, keepdim=False)
-        return cond(eps, 0.0, not_eq, event_dim=event_dim)
-
-    return _consequent_differs
